调整config
