+++ .gitignore
@@ -1,3 +1,4 @@
 go.mod
 /pkg/
 /.idea/
+.DS_Store
