--- resources/App.go
@@ -1,11 +1,0 @@
-package resources
-
-type App struct {
-	Port string
-	//管理员钱包地址
-	AdminAddress string
-	//链地址
-	DeployAddress string
-	//合约地址
-	ContractAddress string
-}
